pyproject.toml
[project]
name = "agentic-text-hackhaton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.1.9",
    "aiohttp>=3.11.13",
    "chainlit>=2.3.0",
    "duckdb>=1.2.1",
    "fastapi[standard]>=0.115.11",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.19",
    "openai>=1.65.5",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.13.3",
]
